Split `charts/` module into `single_category_charts/` so that we can implement `dual_category_charts/` alongside
