+++ metrics/api/serializers/charts/single_category_charts.py
@@ -5,6 +5,7 @@ from rest_framework import serializers
 from rest_framework.request import Request
 
 from metrics.api.serializers import help_texts, plots
+from metrics.api.serializers.charts.common import BaseChartsSerializer
 from metrics.domain.charts.colour_scheme import RGBAChartLineColours
 from metrics.domain.charts.common_charts.plots.line_multi_coloured.properties import (
     ChartLineTypes,
@@ -14,7 +15,6 @@ from metrics.domain.common.utils import (
     DEFAULT_CHART_WIDTH,
     DEFAULT_X_AXIS,
     DEFAULT_Y_AXIS,
-    ChartAxisFields,
     ChartTypes,
     DEFAULT_Y_AXIS_MINIMUM_VAlUE,
 )
@@ -71,71 +71,7 @@ class ChartPlotsListSerializer(serializers.ListSerializer):
             super().__init__(*args, **kwargs)
 
 
-FILE_FORMAT_CHOICES: list[str] = ["svg", "png", "jpg", "jpeg"]
-
-
-class ChartsSerializer(serializers.Serializer):
-    file_format = serializers.ChoiceField(
-        choices=FILE_FORMAT_CHOICES,
-        help_text=help_texts.CHART_FILE_FORMAT_FIELD,
-        default="svg",
-    )
-    chart_height = serializers.IntegerField(
-        help_text=help_texts.CHART_HEIGHT,
-        default=DEFAULT_CHART_HEIGHT,
-        allow_null=True,
-    )
-    chart_width = serializers.IntegerField(
-        help_text=help_texts.CHART_WIDTH,
-        default=DEFAULT_CHART_WIDTH,
-        allow_null=True,
-    )
-    x_axis = serializers.ChoiceField(
-        choices=ChartAxisFields.choices(),
-        required=False,
-        allow_blank=True,
-        allow_null=True,
-        help_text=help_texts.CHART_X_AXIS,
-        default=DEFAULT_X_AXIS,
-    )
-    x_axis_title = serializers.CharField(
-        required=False,
-        allow_blank=True,
-        allow_null=True,
-        default="",
-        help_text=help_texts.CHART_X_AXIS_TITLE,
-    )
-    y_axis = serializers.ChoiceField(
-        choices=ChartAxisFields.choices(),
-        required=False,
-        allow_blank=True,
-        allow_null=True,
-        help_text=help_texts.CHART_Y_AXIS,
-        default=DEFAULT_Y_AXIS,
-    )
-    y_axis_title = serializers.CharField(
-        required=False,
-        allow_blank=True,
-        allow_null=True,
-        default="",
-        help_text=help_texts.CHART_Y_AXIS_TITLE,
-    )
-    y_axis_minimum_value = serializers.DecimalField(
-        required=False,
-        allow_null=True,
-        help_text=help_texts.CHART_Y_AXIS_MINIMUM_VALUE,
-        default=None,
-        max_digits=10,
-        decimal_places=1,
-    )
-    y_axis_maximum_value = serializers.DecimalField(
-        required=False,
-        allow_null=True,
-        help_text=help_texts.CHART_Y_AXIS_MAXIMUM_VALUE,
-        default=None,
-        max_digits=10,
-        decimal_places=1,
-    )
+class ChartsSerializer(BaseChartsSerializer):
 
     plots = ChartPlotsListSerializer()
