+++ metrics/api/views/charts/__init__.py
@@ -1,0 +1,1 @@
+from .single_category_charts import ChartsView, EncodedChartsView
