+++ metrics/api/views/charts/single_category_charts.py
@@ -18,7 +18,7 @@ from metrics.api.serializers.charts import (
     EncodedChartsRequestSerializer,
 )
 from metrics.domain.models import ChartRequestParams
-from metrics.interfaces.charts import access
+from metrics.interfaces.charts.single_category_charts import access
 from metrics.interfaces.plots.access import (
     DataNotFoundForAnyPlotError,
     InvalidPlotParametersError,
