+++ metrics/interfaces/charts/single_category_charts/access.py
@@ -23,7 +23,7 @@ from metrics.domain.models import (
     PlotGenerationData,
 )
 from metrics.domain.models.plots_text import PlotsText
-from metrics.interfaces.charts import calculations
+from metrics.interfaces.charts.single_category_charts import calculations
 from metrics.interfaces.plots.access import PlotsInterface
 from metrics.utils.type_hints import CORE_MODEL_MANAGER_TYPE
