+++ metrics/interfaces/charts/single_category_charts/calculations.py
