+++ tests/unit/metrics/api/serializers/charts/test_single_category_charts.py
