+++ tests/unit/metrics/api/views/charts/test_single_category_charts.py
